+++ AmplifyPlugins/Storage/Tests/AWSS3StoragePluginTests/Request/AWSS3StorageGetDataRequestTests.swift
@@ -6,7 +6,7 @@
 //
 
 import XCTest
-import Amplify
+@testable import Amplify
 @testable import AWSS3StoragePlugin
 
 class StorageDownloadDataRequestTests: XCTestCase {
@@ -93,5 +93,21 @@ class StorageDownloadDataRequestTests: XCTestCase {
         XCTAssertEqual(field, StorageErrorConstants.keyIsEmpty.field)
         XCTAssertEqual(description, StorageErrorConstants.keyIsEmpty.errorDescription)
         XCTAssertEqual(recovery, StorageErrorConstants.keyIsEmpty.recoverySuggestion)
+    }
+
+    /// Given: StorageDownloadDataRequest with an invalid StringStoragePath
+    /// When: Request validation is executed
+    /// Then: There is no error returned even though the storage path is invalid
+    /// There is no error because the path validation is done at operation execution time and not part of the request
+    func testValidateWithStoragePath() {
+        let options = StorageDownloadDataRequest.Options(accessLevel: .private,
+                                                    targetIdentityId: "",
+                                                    pluginOptions: testPluginOptions)
+        let path = StringStoragePath(resolve: { input in return "my/path/"})
+        let request = StorageDownloadDataRequest(path: path, options: options)
+
+        let storageErrorOptional = request.validate()
+
+        XCTAssertNil(storageErrorOptional)
     }
 }