+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 2.31.1 (2024-04-26)
+
+### Bug Fixes
+
+- **storage**: retrieve accesslevel before storage service (#3641)
+
+## 2.31.0 (2024-04-26)
+
+### Features
+
+- **api**: add authorizationMode to GraphQLRequest (#3630)
+
+## 2.30.0 (2024-04-26)
+
+### Features
+
+- **all**: Configure plugins with AmplifyOutputs (#3567)
+
 ## 2.29.3 (2024-04-22)
 
 ### Bug Fixes